// frontend/src/interfaces/EventType.ts
export enum EventType {
  CommentCreated = 'comment-created',
  CardMoved = 'card-moved',
  LaneAmountChanged = 'lane-amount-changed',
  AmountChanged = 'amount-changed',
  AttributeChanged = 'attribute-changed',
  NameChanged = 'name-changed',
  ClosedAtChanged = 'closed-at-changed',
  NextFollowUpAtChanged = 'next-follow-up-at-changed',
  NextFollowUpAtWarning = 'next-follow-up-at-warning',
  Created = 'created',
  Assigned = 'assigned',
  Deleted = 'deleted',
}